+++ src/main/config/configparser_test.go
@@ -26,5 +26,5 @@ func TestParseConfigValid(t *testing.T) {
 	assert.Len(t, config.Schedules, 1)
 	assert.Equal(t, "* * * * *", config.Schedules[0].Cron)
 	assert.Equal(t, "XXBTZEUR", config.Schedules[0].Pair)
-	assert.Equal(t, float32(40.00), config.Schedules[0].Amount)
+	assert.Equal(t, 40.00, config.Schedules[0].Amount)
 }
