+++ src/main/scheduler/model/order_test.go
@@ -9,5 +9,5 @@ import (
 func TestOrder_Amount(t *testing.T) {
 	order := NewOrder("testpair", 1000.0, 500.0)
 
-	assert.Equal(t, order.Amount(), float32(0.5))
+	assert.Equal(t, order.Amount(), 0.5)
 }
